generic integration tests
